refactor: Computer 클래스에서 화면 렌더링 관련 메서드 Dom 클래스 만들어 분리
