+++ src/Computer.js
@@ -1,11 +1,6 @@
 export default class Computer {
   constructor(NUMBER_LENGTH) {
     this.NUMBER_LENGTH = NUMBER_LENGTH;
-    this.resultArea = document.getElementById("result");
-  }
-
-  clearResultArea() {
-    this.resultArea.innerText = "";
   }
 
   makeNumbers() {
@@ -18,20 +13,5 @@ export default class Computer {
 
   changeStringToNumberArray(string) {
     return [...string].map(value => Number(value));
-  }
-
-  createNodeParentChild(parentTag, childTag, childText) {
-    const parent = document.createElement(parentTag);
-    const element = document.createElement(childTag);
-    element.innerText = childText;
-    parent.append(element);
-    return parent;
-  }
-
-  createButton(text, id) {
-    const button = document.createElement("button");
-    button.innerText = text;
-    button.id = id;
-    return button;
   }
 }
